refactor(video-card): add avatar component
