+++ src/components/VideoCard/types.ts
@@ -1,0 +1,54 @@
+import type { ThreePointV2 } from '~/models/video/appForYou'
+
+export interface Video {
+  id: number
+  duration?: number
+  durationStr?: string
+  title: string
+  desc?: string
+  cover: string
+
+  authorList?: Author[]
+  view?: number
+  viewStr?: string
+  danmaku?: number
+  danmakuStr?: string
+
+  publishedTimestamp?: number
+  capsuleText?: string
+
+  bvid?: string
+  aid?: number
+  // used for live
+  roomid?: number
+  epid?: number
+  goto?: string
+  /** After set the `url`, clicking the video will navigate to this url. It won't be affected by aid, bvid or epid */
+  url?: string
+  /** If you want to show preview video, you should set the cid value */
+  cid?: number
+
+  followed?: boolean
+  liveStatus?: number
+
+  tag?: string
+  rank?: number
+  type?: 'horizontal' | 'vertical' | 'bangumi'
+  threePointV2: ThreePointV2[]
+
+  badge?: {
+    bgColor: string
+    color: string
+    iconUrl?: string
+    text: string
+  }
+}
+
+export interface Author {
+  author?: string
+  /** After set the `authorUrl`, clicking the author's name or avatar will navigate to this url. It won't be affected by mid */
+  authorUrl?: string
+  authorFace: string
+  followed?: boolean | undefined
+  mid?: number
+}
